requirements.txt
streamlit==1.30.0
langchain==0.1.0
langchain_groq==0.0.5
python-dotenv==1.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
